requirements.txt
flask==3.0.0
langchain==0.1.0
langgraph==0.0.15
faiss-cpu==1.7.4
python-dotenv==1.0.0
python-docx==1.0.1
PyMuPDF==1.23.8
openpyxl==3.1.2
langchain-community==0.0.13
langchain-core==0.1.12
langchain-ollama==0.0.3
langchain-huggingface==0.0.7
sentence-transformers==2.2.2
